+++ src/main/java/com/ssg/sausagecartsharecalculationapi/cartsharecal/dto/response/CartShareCalDtlFindInfo.java
@@ -1,6 +1,6 @@
-package com.ssg.sausagecartsharecalculationapi.dutchpay.dto.response;
+package com.ssg.sausagecartsharecalculationapi.cartsharecal.dto.response;
 
-import com.ssg.sausagecartsharecalculationapi.dutchpay.entity.CartShareCalDtl;
+import com.ssg.sausagecartsharecalculationapi.cartsharecal.entity.CartShareCalDtl;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
