+++ src/main/java/com/ssg/sausagecartsharecalculationapi/cartsharecal/dto/response/CartShareCalSaveResponse.java
@@ -1,6 +1,6 @@
-package com.ssg.sausagecartsharecalculationapi.dutchpay.dto.response;
+package com.ssg.sausagecartsharecalculationapi.cartsharecal.dto.response;
 
-import com.ssg.sausagecartsharecalculationapi.dutchpay.entity.CartShareCal;
+import com.ssg.sausagecartsharecalculationapi.cartsharecal.entity.CartShareCal;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
