+++ src/test/http/http-client.env.json
@@ -1,5 +1,5 @@
 {
   "local": {
-    "host": "http://localhost:8083"
+    "host": "http://localhost:8084"
   }
 }
